+++ eva/src/components/footer.vue
@@ -16,7 +16,7 @@ export default {
 
   data () {
     return {
-      version: '1.51.0',
+      version: '1.52.0 with-graph',
       color: { },
     } 
   },