+++ eva/src/js/componentsSettings.js
@@ -1,22 +1,24 @@
 
-import {  mdiGraph, mdiGestureTap, mdiBrush, mdiChartLine,   mdiPoll, mdiViewGrid, 
-  mdiCardTextOutline, mdiChartGantt, mdiChartMultiline,  mdiNumeric,  mdiTableLarge,  mdiFormatListBulleted,  mdiCalendarMonth, } from '@mdi/js'
+import {  mdiGraph, mdiGestureTap, mdiBrush, mdiViewGrid, 
+  mdiCardTextOutline, mdiChartGantt, mdiChartMultiline,  mdiNumeric,  mdiTableLarge, mdiChartPie,  mdiFormatListBulleted,  mdiCalendarMonth, mdiGraphql } from '@mdi/js'
 
 export default {  
   tools: [
     {name: 'Таблица', img: mdiTableLarge, type: 'table'},
-    {name: 'Линейный график', img: mdiChartLine, type: 'lineChart'},
+    //{name: 'Линейный график', img: mdiChartLine, type: 'lineChart'},
     {name: 'Мультилинейный график', img: mdiChartMultiline, type: 'multiLine'},
-    {name: 'Cтолбцовая диаграмма', img: mdiPoll , type: 'barchart'},
+    //{name: 'Cтолбцовая диаграмма', img: mdiPoll , type: 'barchart'},
+    {name: 'Круговая диаграмма', img: mdiChartPie, type: 'piechart'},
     {name: 'Выпадающий список', img: mdiFormatListBulleted, type: 'select'},
     {name: 'Выбор времени', img: mdiCalendarMonth, type: 'picker'},
-    {name: 'Граф', img: mdiGraph , type: 'graph'},
+    {name: 'Граф', img: mdiGraphql , type: 'ygraph'},
     {name: 'Диаграмма Гантта', img: mdiChartGantt, type: 'guntt'},
     {name: 'Плитка', img: mdiViewGrid , type: 'tile'},
     {name: 'Изображение', img: mdiBrush , type: 'csvg'},
     {name: 'Число', img: mdiNumeric, type: 'single'},
     {name: 'Кнопка', img: mdiGestureTap , type: 'button'},
     {name: 'Текстовый блок', img: mdiCardTextOutline , type: 'textarea'},
+    {name: 'Граф_old', img: mdiGraph , type: 'graph'},
   ],
   size: {
     picker: {
@@ -27,9 +29,13 @@ export default {
       width: 480,
       height: 205
     },
-    lineChart: {
-      width: 700,
-      height: 400
+    // lineChart: {
+    //   width: 700,
+    //   height: 400
+    // },
+    piechart: {
+      width: 600,
+      height: 550,
     },
     graph: {
       width: 1000,
@@ -59,10 +65,10 @@ export default {
       width: 400,
       height: 80
     },
-    barchart: {
-      width: 700,
-      height: 450
-    },
+    // barchart: {
+    //   width: 700,
+    //   height: 450
+    // },
     tile: {
       width: 470,
       height: 400
@@ -70,12 +76,17 @@ export default {
     csvg: {
       width: 500,
       height: 400
-    }  
+    },
+    ygraph: {
+      width: 930,
+      height: 850
+    },  
   },
   icons: {
     table: mdiTableLarge,
-    lineChart: mdiChartLine,
+    //lineChart: mdiChartLine,
     multiLine: mdiChartMultiline,
+    piechart: mdiChartPie,
     graph: mdiGraph,
     guntt: mdiChartGantt,
     select: mdiFormatListBulleted,
@@ -83,14 +94,16 @@ export default {
     single: mdiNumeric,
     button: mdiGestureTap,
     textarea: mdiCardTextOutline,
-    barchart: mdiPoll,
+  //  barchart: mdiPoll,
     tile: mdiViewGrid,
     csvg: mdiBrush,
+    ygraph: mdiGraphql
   }, 
   options: {
-    lineChart: ['visible','level','boxShadow','lastDot'],
-    multiLine: ['visible','level','boxShadow','united','lastDot','metrics'], 
-    barchart: ['visible','level','boxShadow'],
+    //lineChart: ['visible','level','boxShadow','lastDot'],
+    multiLine: ['visible','level','boxShadow','united','lastDot','metrics', 'timeFormat'], 
+    piechart: ['visible','level','metricsRelation','showlegend','positionlegend', 'colorsPie','themes'],
+    //barchart: ['visible','level','boxShadow'],
     table: ['visible','level','boxShadow','rowcolor','columncolor','cellcolor'],
     select: ['visible','level','boxShadow','multiple'],
     picker: ['visible','level'],
@@ -100,26 +113,31 @@ export default {
     textarea: ['visible','level'],
     guntt: ['visible','level','timeFormat'],
     tile: ['visible','level','widthTile','heightTile'],
-    csvg: ['visible','level','tooltip']
+    csvg: ['visible','level','tooltip'],
+    ygraph: ['visible','level']
   },
-  reporstElements: ['table','lineChart','multiLine','barchart','guntt','tile','csvg'],
+  reporstElements: ['table','multiLine','piechart','guntt','tile','csvg'],
   reports: {
     table: {
       tooltip: 'Таблица',
       icon: mdiTableLarge
     },
-    lineChart: {
-      tooltip: 'Линейный график',
-      icon: mdiChartLine
-    },
+    // lineChart: {
+    //   tooltip: 'Линейный график',
+    //   icon: mdiChartLine
+    // },
     multiLine: {
       tooltip: 'Мультилинейный график',
       icon: mdiChartMultiline
     },
-    barchart: {
-      tooltip: 'Столбцовая диаграмма',
-      icon: mdiPoll
+    piechart: {
+      tooltip: 'Круговая диаграмма',
+      icon: mdiChartPie
     },
+    // barchart: {
+    //   tooltip: 'Столбцовая диаграмма',
+    //   icon: mdiPoll
+    // },
     graph: {
       tooltip: 'Граф',
       icon: mdiGraph
@@ -135,6 +153,10 @@ export default {
     csvg: {
       tooltip: 'Изображение',
       icon: mdiBrush
+    },
+    ygraph: {
+      tooltip: 'Граф',
+      icon: mdiGraphql
     },
   }
 }